+++ pkg/stanza/operator/input/windows/buffer.go
@@ -14,7 +14,7 @@
 
 // +build windows
 
-package windows // import "github.com/open-telemetry/opentelemetry-log-collection/operator/input/windows"
+package windows // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/windows"
 
 import (
 	"bytes"