+++ receiver/kubeletstatsreceiver/internal/kubelet/metadata.go
@@ -23,6 +23,8 @@ import (
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	stats "k8s.io/kubelet/pkg/apis/stats/v1alpha1"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/kubeletstatsreceiver/internal/metadata"
 )
 
 type MetadataLabel string
@@ -55,18 +57,18 @@ func ValidateMetadataLabelsConfig(labels []MetadataLabel) error {
 }
 
 type Metadata struct {
-	Labels                  map[MetadataLabel]bool
-	PodsMetadata            *v1.PodList
-	DetailedPVCLabelsSetter func(volCacheID, volumeClaim, namespace string, labels map[string]string) error
+	Labels                    map[MetadataLabel]bool
+	PodsMetadata              *v1.PodList
+	DetailedPVCResourceGetter func(volCacheID, volumeClaim, namespace string) ([]metadata.ResourceMetricsOption, error)
 }
 
 func NewMetadata(
 	labels []MetadataLabel, podsMetadata *v1.PodList,
-	detailedPVCLabelsSetter func(volCacheID, volumeClaim, namespace string, labels map[string]string) error) Metadata {
+	detailedPVCResourceGetter func(volCacheID, volumeClaim, namespace string) ([]metadata.ResourceMetricsOption, error)) Metadata {
 	return Metadata{
-		Labels:                  getLabelsMap(labels),
-		PodsMetadata:            podsMetadata,
-		DetailedPVCLabelsSetter: detailedPVCLabelsSetter,
+		Labels:                    getLabelsMap(labels),
+		PodsMetadata:              podsMetadata,
+		DetailedPVCResourceGetter: detailedPVCResourceGetter,
 	}
 }
 
@@ -78,45 +80,46 @@ func getLabelsMap(metadataLabels []MetadataLabel) map[MetadataLabel]bool {
 	return out
 }
 
-// setExtraLabels sets extra labels in `labels` map based on provided metadata label.
-func (m *Metadata) setExtraLabels(
-	labels map[string]string, podRef stats.PodReference,
-	extraMetadataLabel MetadataLabel, extraMetadataFrom string) error {
+// getExtraResources gets extra resources based on provided metadata label.
+func (m *Metadata) getExtraResources(podRef stats.PodReference, extraMetadataLabel MetadataLabel,
+	extraMetadataFrom string) ([]metadata.ResourceMetricsOption, error) {
 	// Ensure MetadataLabel exists before proceeding.
 	if !m.Labels[extraMetadataLabel] || len(m.Labels) == 0 {
-		return nil
+		return nil, nil
 	}
 
 	// Cannot proceed, if metadata is unavailable.
 	if m.PodsMetadata == nil {
-		return errors.New("pods metadata were not fetched")
+		return nil, errors.New("pods metadata were not fetched")
 	}
 
 	switch extraMetadataLabel {
 	case MetadataLabelContainerID:
 		containerID, err := m.getContainerID(podRef.UID, extraMetadataFrom)
 		if err != nil {
-			return err
+			return nil, err
 		}
-		labels[conventions.AttributeContainerID] = containerID
+		return []metadata.ResourceMetricsOption{metadata.WithContainerID(containerID)}, nil
 	case MetadataLabelVolumeType:
 		volume, err := m.getPodVolume(podRef.UID, extraMetadataFrom)
 		if err != nil {
-			return err
+			return nil, err
 		}
 
-		getLabelsFromVolume(volume, labels)
+		ro := getResourcesFromVolume(volume)
 
 		// Get more labels from PersistentVolumeClaim volume type.
 		if volume.PersistentVolumeClaim != nil {
 			volCacheID := fmt.Sprintf("%s/%s", podRef.UID, extraMetadataFrom)
-			if err := m.DetailedPVCLabelsSetter(volCacheID, labels[labelPersistentVolumeClaimName], podRef.Namespace,
-				labels); err != nil {
-				return fmt.Errorf("failed to set labels from volume claim: %w", err)
+			pvcResources, err := m.DetailedPVCResourceGetter(volCacheID, volume.PersistentVolumeClaim.ClaimName, podRef.Namespace)
+			if err != nil {
+				return nil, fmt.Errorf("failed to set labels from volume claim: %w", err)
 			}
+			ro = append(ro, pvcResources...)
 		}
+		return ro, nil
 	}
-	return nil
+	return nil, nil
 }
 
 // getContainerID retrieves container id from metadata for given pod UID and container name,